requirements.txt
openai
faiss-cpu
sentence-transformers
pandas
numpy
matplotlib
pyyaml
tqdm